+++ packages/designer/src/components/widgets/project-config/index.vue
@@ -12,7 +12,10 @@
         label="Logo"
         name="logo"
         :editor="FileSetter"
-        :props="{ attachment: { accept: '.png,.jpg,.jpeg' } }"></XField>
+        :props="{
+          attachment: { accept: '.png,.jpg,.jpeg,.gif,.svg' },
+          acceptFilter: true
+        }"></XField>
       <XField label="主题切换" name="themeSwitchable" editor="switch"></XField>
       <!-- <XField label="模拟数据" name="mock" editor="switch"></XField> -->
     </XForm>