+++ packages/materials/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.13.10](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.13.9...@vtj/materials@0.13.10) (2025-08-16)
+
+
+### Bug Fixes
+
+* 🐛 antd物料slot问题 ([d8c1f9b](https://gitee.com/newgateway/vtj/commits/d8c1f9b1e4cbd566449ecf27d64482c94ecc08ba))
+
+
+
+
+
+## [0.13.9](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.13.8...@vtj/materials@0.13.9) (2025-08-14)
+
+**Note:** Version bump only for package @vtj/materials
+
+
+
+
+
+## [0.13.8](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.13.7...@vtj/materials@0.13.8) (2025-08-13)
+
+**Note:** Version bump only for package @vtj/materials
+
+
+
+
+
+## [0.13.7](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.13.6...@vtj/materials@0.13.7) (2025-08-08)
+
+**Note:** Version bump only for package @vtj/materials
+
+
+
+
+
 ## [0.13.6](https://gitee.com/newgateway/vtj/compare/@vtj/materials@0.13.5...@vtj/materials@0.13.6) (2025-08-08)
 
 **Note:** Version bump only for package @vtj/materials